renaming redux context to redux store context
