+++ src/components/context/index.ts
@@ -1,1 +1,1 @@
-export * from './Redux';
+export * from './ReduxStore';
